change how to deal with process
