+++ srcs/execute/execute_command_struct.c
@@ -6,7 +6,7 @@
 /*   By: sosugimo <sosugimo@student.42tokyo.>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/27 12:38:26 by sosugimo          #+#    #+#             */
-/*   Updated: 2022/01/27 12:38:41 by sosugimo         ###   ########.fr       */
+/*   Updated: 2022/01/27 13:24:59 by sosugimo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
